Move tensors package to root
