+++ AGENTS/CODEBASE_REGISTRY.md
@@ -8,5 +8,6 @@ This document lists the independent project directories maintained in this repos
 - **time_sync** — system clock offset helpers.
 - **AGENTS/tools** — shared helper scripts for repository management.
 - **fontmapper** — ASCII rendering and font mapping experiments.
+- **tensors** — tensor backends and abstraction layer.
 
 Additional codebases can be added here as the repository grows.
