+++ AGENTS/experience_reports/1749605172_v1_Tensors_Split.md
@@ -1,0 +1,8 @@
+# Tensors Package Split
+
+## Overview
+Moved `speaktome/tensors` into a standalone `tensors` codebase and updated
+imports across the project.
+
+## Prompts
+- User: "pull speaktome/tensors into its own root folder and codebase registration and go through the entire testing and speaktome codebase to fix references to it"
