+++ speaktome/core/beam_graph_operator.py
@@ -6,7 +6,7 @@ try:
 
     import torch
 
-    from ..tensors import (
+    from tensors import (
         AbstractTensorOperations,
         get_tensor_operations,
     )
