+++ tensor printing/tensor_printing/press.py
@@ -4,7 +4,7 @@
 from __future__ import annotations
 
 from typing import Any
-from speaktome.tensors.abstraction import AbstractTensorOperations
+from tensors.abstraction import AbstractTensorOperations
 
 from .ruler import Ruler
