+++ tensors/AGENTS.md
