+++ tensors/README.md
@@ -1,0 +1,5 @@
+# Tensors Package
+
+This package provides an abstract tensor interface and backend implementations.
+It was moved out of the main `speaktome` project to serve as a standalone
+codebase. Optional extras let you install JAX, C, and NumPy backends.
