+++ tensors/pyproject.toml
@@ -1,0 +1,32 @@
+[build-system]
+requires = ["setuptools>=61", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "tensors"
+version = "0.1.0"
+description = "Tensor backends and abstraction layer"
+readme = "README.md"
+requires-python = ">=3.8"
+license = { text = "MIT" }
+authors = [{ name = "Tensors Authors" }]
+
+[project.optional-dependencies]
+jax = [
+  "jax[cpu]>=0.4"
+]
+ctensor = [
+  "cffi>=1.15",
+  "setuptools>=61",
+  "ziglang>=0.14"
+]
+numpy = [
+  "numpy>=1.26"
+]
+dev = [
+  "pytest>=8.0"
+]
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["tensors*"]
